ref: Move Github integration into standalone module
