--- server/polar/api/__init__.py
@@ -1,6 +1,0 @@
-from fastapi import APIRouter
-
-from polar.api.v1 import router as v1_router
-
-router = APIRouter(prefix="/api")
-router.include_router(v1_router)
