+++ server/polar/app.py
@@ -1,12 +1,12 @@
 import structlog
 from fastapi import FastAPI
 from fastapi.routing import APIRoute
-from polar.api import router
-from polar.config import settings
-from polar.logging import configure as configure_logging
 from starlette.middleware.cors import CORSMiddleware
 
 from polar import receivers  # noqa
+from polar.api.endpoints import router
+from polar.config import settings
+from polar.logging import configure as configure_logging
 
 log = structlog.get_logger()
