+++ server/polar/integrations/github/actions/organization.py
@@ -1,0 +1,128 @@
+from datetime import datetime
+
+import structlog
+
+from polar.actions.organization import OrganizationActions
+from polar.models import Organization, User
+from polar.platforms import Platforms
+from polar.postgres import AsyncSession
+from polar.schema.organization import OrganizationCreate
+
+from .. import client as github
+from .repository import github_repository
+
+log = structlog.get_logger()
+
+
+class GithubOrganization(OrganizationActions):
+    async def get_by_external_id(
+        self, session: AsyncSession, external_id: int
+    ) -> Organization | None:
+        return await self.get_by_platform(session, Platforms.github, external_id)
+
+    async def fetch_installations(
+        self, session: AsyncSession, user: User
+    ) -> list[OrganizationCreate] | None:
+        oauth = user.get_platform_oauth_account(Platforms.github)
+        if not oauth:
+            # TODO Handle
+            return None
+
+        client = github.get_client(oauth.access_token)
+        response = (
+            await client.rest.apps.async_list_installations_for_authenticated_user()
+        )
+        github.ensure_expected_response(response)
+
+        installations = response.parsed_data.installations
+        log.debug(
+            "github.installations.fetch.success",
+            user_id=user.id,
+            installation_count=len(installations),
+        )
+        if not installations:
+            return None
+
+        return [OrganizationCreate.from_github_installation(i) for i in installations]
+
+    async def install(
+        self, session: AsyncSession, user: User, installation_id: int
+    ) -> Organization | None:
+        installations = await self.fetch_installations(session, user)
+        if not installations:
+            return None
+
+        # Ideally, we could fetch the specific resource with /apps/installation/{id}
+        # instead. However, Github only provides the installation_id and no verification
+        # token. Therefore, using it would expose us to CSRF risks, e.g malicious user
+        # guessing other installation IDs to get connected to them.
+        filtered = [i for i in installations if i.installation_id == installation_id]
+        if not filtered:
+            return None
+
+        to_create = filtered.pop()
+        organization = await self.upsert(session, to_create)
+        if not organization:
+            return None
+
+        # TODO: Better error handling?
+        await self.add_user(session, organization, user)
+        repositories = await github_repository.install_for_organization(
+            session, organization, installation_id
+        )
+        # TODO: Use SQLAlchemy features here vs. hard setter
+        organization.repositories = repositories
+        return organization
+
+    async def suspend(
+        self,
+        session: AsyncSession,
+        installation_id: int,
+        suspended_by: int,
+        suspended_at: datetime | None = None,
+        external_user_id: int | None = None,
+    ) -> bool:
+        org = await self.get_by(session, installation_id=installation_id)
+        if not org:
+            return False
+
+        if suspended_at is None:
+            suspended_at = datetime.utcnow()
+
+        # TODO: Return object instead?
+        await org.update(
+            session,
+            installation_suspended_at=suspended_at,
+            status=Organization.Status.SUSPENDED,
+            installation_suspended_by=suspended_by,
+            installation_suspender=external_user_id,
+        )
+        return True
+
+    async def unsuspend(
+        self,
+        session: AsyncSession,
+        installation_id: int,
+        external_user_id: int | None = None,
+    ) -> bool:
+        org = await self.get_by(session, installation_id=installation_id)
+        if not org:
+            return False
+
+        # TODO: Return object instead?
+        await org.update(
+            session,
+            installation_suspended_at=None,
+            status=Organization.Status.ACTIVE,
+            installation_suspended_by=None,
+            installation_suspender=external_user_id,
+        )
+        return True
+
+    async def remove(self, session: AsyncSession, installation_id: int) -> bool:
+        # TODO: Add security re: installation ownership?
+        # TODO: Soft deletes at least OR even versioning
+        return await self.delete(session, installation_id=installation_id)
+
+
+github_organization = GithubOrganization(Organization)
