+++ server/polar/integrations/github/tasks/webhook.py
@@ -3,8 +3,7 @@ from typing import Any
 
 import structlog
 
-from polar import actions
-from polar.clients import github
+from polar.integrations.github import client as github
 from polar.models import Issue, Organization, PullRequest, Repository
 from polar.platforms import Platforms
 from polar.postgres import AsyncSession
@@ -13,6 +12,8 @@ from polar.schema.organization import OrganizationCreate
 from polar.schema.pull_request import FullPullRequestCreate, PullRequestRead
 from polar.schema.repository import RepositoryCreate
 from polar.worker import get_db_session, sync_worker, task
+
+from .. import actions
 
 log = structlog.get_logger()
