+++ server/polar/tasks.py
@@ -1,0 +1,3 @@
+from polar.integrations.github import tasks as github
+
+__all__ = ["github"]
