--- server/polar/tasks/__init__.py
@@ -1,3 +1,0 @@
-from polar.tasks import github
-
-__all__ = ["github"]
