--- server/polar/tasks/github/__init__.py
@@ -1,3 +1,0 @@
-from polar.tasks.github import webhook
-
-__all__ = ["webhook"]
