+++ server/tests/integrations/github/test_github_webhooks.py
@@ -4,8 +4,8 @@ from typing import Any
 
 import pytest
 
-from polar import actions
-from polar.clients import github
+from polar.integrations.github import actions
+from polar.integrations.github import client as github
 from polar.models.organization import Organization
 from polar.platforms import Platforms
 from polar.postgres import AsyncSession, AsyncSessionLocal
